Sync Fork from Upstream Repo

+++ go/libraries/doltcore/schema/alterschema/dropcolumn.go
@@ -21,8 +21,9 @@ import (
 
 	"github.com/dolthub/dolt/go/libraries/doltcore/doltdb"
 	"github.com/dolthub/dolt/go/libraries/doltcore/schema"
-	"github.com/dolthub/dolt/go/libraries/doltcore/schema/encoding"
 )
+
+var ErrKeylessAltTbl = errors.New("schema alterations not supported for keyless tables")
 
 // DropColumn drops a column from a table, and removes its associated cell values
 func DropColumn(ctx context.Context, tbl *doltdb.Table, colName string, foreignKeys []doltdb.ForeignKey) (*doltdb.Table, error) {
@@ -30,16 +31,17 @@ func DropColumn(ctx context.Context, tbl *doltdb.Table, colName string, foreignK
 		panic("invalid parameters")
 	}
 
-	tblSch, err := tbl.GetSchema(ctx)
-
+	sch, err := tbl.GetSchema(ctx)
 	if err != nil {
 		return nil, err
 	}
 
-	allCols := tblSch.GetAllCols()
+	if schema.IsKeyless(sch) {
+		return nil, ErrKeylessAltTbl
+	}
 
 	var dropTag uint64
-	if col, ok := allCols.GetByName(colName); !ok {
+	if col, ok := sch.GetAllCols().GetByName(colName); !ok {
 		return nil, schema.ErrColNotFound
 	} else if col.IsPartOfPK {
 		return nil, errors.New("Cannot drop column in primary key")
@@ -60,8 +62,8 @@ func DropColumn(ctx context.Context, tbl *doltdb.Table, colName string, foreignK
 		}
 	}
 
-	for _, index := range tblSch.Indexes().IndexesWithColumn(colName) {
-		_, err = tblSch.Indexes().RemoveIndex(index.Name())
+	for _, index := range sch.Indexes().IndexesWithColumn(colName) {
+		_, err = sch.Indexes().RemoveIndex(index.Name())
 		if err != nil {
 			return nil, err
 		}
@@ -72,7 +74,7 @@ func DropColumn(ctx context.Context, tbl *doltdb.Table, colName string, foreignK
 	}
 
 	cols := make([]schema.Column, 0)
-	err = allCols.Iter(func(tag uint64, col schema.Column) (stop bool, err error) {
+	err = sch.GetAllCols().Iter(func(tag uint64, col schema.Column) (stop bool, err error) {
 		if col.Name != colName {
 			cols = append(cols, col)
 		}
@@ -83,35 +85,12 @@ func DropColumn(ctx context.Context, tbl *doltdb.Table, colName string, foreignK
 		return nil, err
 	}
 
-	colColl, err := schema.NewColCollection(cols...)
-	if err != nil {
-		return nil, err
-	}
-
+	colColl := schema.NewColCollection(cols...)
 	newSch, err := schema.SchemaFromCols(colColl)
 	if err != nil {
 		return nil, err
 	}
-	newSch.Indexes().AddIndex(tblSch.Indexes().AllIndexes()...)
+	newSch.Indexes().AddIndex(sch.Indexes().AllIndexes()...)
 
-	vrw := tbl.ValueReadWriter()
-	schemaVal, err := encoding.MarshalSchemaAsNomsValue(ctx, vrw, newSch)
-
-	if err != nil {
-		return nil, err
-	}
-
-	rd, err := tbl.GetRowData(ctx)
-
-	indexData, err := tbl.GetIndexData(ctx)
-	if err != nil {
-		return nil, err
-	}
-	newTable, err := doltdb.NewTable(ctx, vrw, schemaVal, rd, indexData)
-
-	if err != nil {
-		return nil, err
-	}
-
-	return newTable, nil
+	return tbl.UpdateSchema(ctx, newSch)
 }
